MAC-72: Type provider macros as float so OpenAI can compile the schema (#58)
* MAC-72: Type provider macros as float so OpenAI can compile the schema

The request goes to OpenAI with "strict": true. In strict mode OpenAI
compiles the JSON schema into a decoder grammar before it generates a
token.

Pydantic renders a Decimal field as a union of a number and a patterned
string, and that pattern holds a negative lookahead. OpenAI cannot
compile that shape. Generation never starts. The call returns
status=incomplete with reason=max_output_tokens and zero tokens used,
which points at a budget that was never the problem. Raising the budget
from 4096 to 16000 changes nothing.

Type calories, protein_g, and fiber_g as float. The same call then
returns 24 items. Precision is unchanged: _rounded_macro already
converts with Decimal(str(value)), and str() on a float returns the
shortest string that round-trips.

Add a test that walks the generated schema and fails on any property
declaring anyOf or oneOf. It asserts on structure, not on the
serialised text, because pydantic copies a class docstring into
description and a keyword search matches prose.


* MAC-72: Keep the provider model's notes out of the schema

Pydantic copies a class docstring into the JSON schema's description, and
the SDK sends that schema on every analysis call. The 896-character
docstring added in the previous commit was more than half the request
schema. OpenAI was billed for it as input tokens and read it as
instructions about the task, though it only discussed Python types and
decoder grammars.

Move the notes to a comment above the class. The schema drops from 1736
to 806 characters.

Add a test asserting the schema carries no description. The first
version of the union test searched the serialised schema for a keyword
and tripped over this same docstring. Fixing the test was the wrong
response. The prose did not belong in the schema at all.


---------

+++ apps/api/ai/provider.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from decimal import Decimal
 from typing import Any
 
 from django.conf import settings
@@ -7,14 +6,34 @@ from openai import OpenAI
 from pydantic import BaseModel, ConfigDict, Field
 
 
+# This model is the wire format OpenAI must generate, not the shape this app
+# stores. Note the comment, not a docstring: pydantic copies a class docstring
+# into the schema's `description`, and the SDK sends the schema on every call.
+# Engineering notes there would be billed as input tokens and read by the model
+# as instructions.
+#
+# The macros are `float` and not `Decimal` on purpose. Pydantic renders a
+# `Decimal` field as `anyOf: [number, string-with-pattern]`, and that pattern
+# holds a negative lookahead. OpenAI compiles this schema into a decoder grammar
+# before it generates anything, and it cannot compile that shape. The call then
+# returns `status=incomplete` with `reason=max_output_tokens` and zero tokens
+# used, which points at a budget that was never the problem.
+#
+# Precision is not lost. `_rounded_macro` in services.py converts each value
+# with `Decimal(str(value))` before it rounds and sums, and `str()` on a float
+# returns the shortest string that round-trips.
+#
+# The wrong choice: reach for `Decimal` here because the database column is a
+# decimal. A provider schema is bound by what a third party can generate, not by
+# how this app stores the result.
 class ProviderFoodItem(BaseModel):
     model_config = ConfigDict(extra="forbid")
 
     name: str = Field(min_length=1, max_length=200)
     portion: str = Field(min_length=1, max_length=100)
-    calories: Decimal = Field(ge=0)
-    protein_g: Decimal = Field(ge=0)
-    fiber_g: Decimal = Field(ge=0)
+    calories: float = Field(ge=0)
+    protein_g: float = Field(ge=0)
+    fiber_g: float = Field(ge=0)
 
 
 class ProviderFoodAnalysis(BaseModel):
